+++ Backend/src/main/resources/db/migration/V3__Create_Punonjesit_Table.sql
@@ -1,5 +1,5 @@
 CREATE TABLE Punonjesit (
-    punonjes_id INT IDENTITY(1,1) PRIMARY KEY,
+    punonjes_id BIGINT IDENTITY(1,1) PRIMARY KEY,
     emri NVARCHAR(255) NOT NULL,
     mbiemri NVARCHAR(255) NOT NULL,
     roli NVARCHAR(100),
