+++ Backend/src/main/resources/db/migration/V8__Create_Furnitoret_Table.sql
@@ -1,5 +1,5 @@
 CREATE TABLE Furnitoret (
-    furnitor_id INT IDENTITY(1,1) PRIMARY KEY,
+    furnitor_id BIGINT IDENTITY(1,1) PRIMARY KEY,
     emri_kompanise NVARCHAR(100) NOT NULL,
     personi_kontaktit NVARCHAR(100),
     email NVARCHAR(100),
